@@ -18,77 +18,79 @@
   "private": true,
   "dependencies": {
     "@agm/core": "^1.1.0",
-    "@angular/animations": "11.2.3",
-    "@angular/cdk": "10.2.4",
-    "@angular/common": "^11.2.3",
-    "@angular/compiler": "11.2.3",
-    "@angular/core": "^11.2.3",
-    "@angular/forms": "11.2.3",
+    "@angular/animations": "18.0.1",
+    "@angular/cdk": "18.0.1",
+    "@angular/common": "^18.0.1",
+    "@angular/compiler": "18.0.1",
+    "@angular/core": "^18.0.1",
+    "@angular/forms": "18.0.1",
     "@angular/http": "^7.2.16",
-    "@angular/localize": "^11.2.3",
-    "@angular/material": "10.2.4",
-    "@angular/platform-browser": "11.2.3",
-    "@angular/platform-browser-dynamic": "11.2.3",
-    "@angular/platform-server": "11.2.3",
-    "@angular/router": "11.2.3",
-    "@auth0/angular-jwt": "^5.0.2",
-    "@ng-bootstrap/ng-bootstrap": "^8.0.0",
-    "@ngx-translate/core": "^13.0.0",
-    "@ngx-translate/http-loader": "^6.0.0",
-    "@sweetalert2/ngx-sweetalert2": "^9.0.0",
-    "ajv": "6.12.5",
-    "alertifyjs": "^1.13.1",
+    "@angular/localize": "^18.0.1",
+    "@angular/material": "18.0.1",
+    "@angular/platform-browser": "18.0.1",
+    "@angular/platform-browser-dynamic": "18.0.1",
+    "@angular/platform-server": "18.0.1",
+    "@angular/router": "18.0.1",
+    "@auth0/angular-jwt": "^5.2.0",
+    "@ng-bootstrap/ng-bootstrap": "^16.0.0",
+    "@ngx-translate/core": "^15.0.0",
+    "@ngx-translate/http-loader": "^8.0.0",
+    "@sweetalert2/ngx-sweetalert2": "^12.3.0",
+    "ajv": "8.16.0",
+    "alertifyjs": "^1.14.0",
     "arrive": "2.4.1",
-    "bootstrap": "4.5.2",
+    "bootstrap": "^5.3.3",
     "bootstrap-material-design": "4.1.3",
     "bootstrap-notify": "3.1.3",
+    "bootstrap-scss": "^5.3.3",
     "bootstrap-select": "^1.13.18",
-    "chartist": "0.11.4",
+    "chartist": "1.3.0",
     "classlist.js": "1.1.20150312",
-    "core-js": "3.6.5",
-    "datatables.net": "^1.10.20",
-    "datatables.net-dt": "^1.10.20",
-    "eslint": "^7.15.0",
-    "express": "4.17.1",
+    "core-js": "3.37.1",
+    "datatables.net": "^2.0.8",
+    "datatables.net-dt": "^2.0.8",
+    "eslint": "^9.4.0",
+    "express": "4.19.2",
     "font-awesome": "^4.7.0",
-    "googleapis": "61.0.0",
+    "googleapis": "139.0.0",
     "hammerjs": "2.0.8",
-    "jquery": "3.5.1",
-    "moment": "2.29.1",
-    "ng-multiselect-dropdown": "^0.2.12",
-    "perfect-scrollbar": "1.5.0",
+    "jquery": "3.7.1",
+    "moment": "2.30.1",
+    "ng-multiselect-dropdown": "^1.0.0",
+    "perfect-scrollbar": "1.5.5",
     "popper.js": "^1.16.1",
-    "rxjs": "6.6.3",
-    "rxjs-compat": "6.6.3",
-    "sweetalert2": "^10.14.0",
+    "rxjs": "7.8.1",
+    "rxjs-compat": "6.6.7",
+    "sweetalert2": "^11.11.1",
     "web-animations-js": "2.3.2",
-    "zone.js": "0.11.4"
+    "zone.js": "0.14.6"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1102.2",
-    "@angular/cli": "^11.2.2",
-    "@angular/compiler-cli": "11.2.3",
-    "@angular/language-service": "11.2.3",
-    "@fortawesome/fontawesome-free": "^5.15.1",
-    "@types/bootstrap": "4.5.0",
-    "@types/chartist": "0.11.0",
-    "@types/datatables.net": "^1.10.18",
+    "@angular-devkit/build-angular": "^18.0.2",
+    "@angular/cli": "^18.0.2",
+    "@angular/compiler-cli": "18.0.1",
+    "@angular/language-service": "18.0.1",
+    "@fortawesome/fontawesome-free": "^6.5.2",
+    "@types/bootstrap": "5.2.10",
+    "@types/chartist": "0.11.1",
+    "@types/datatables.net": "^1.12.0",
+    "@types/datatables.net-select": "^1.3.0",
     "@types/googlemaps": "3.39.14",
-    "@types/jasmine": "3.5.14",
-    "@types/jquery": "3.5.2",
-    "@types/node": "14.11.5",
-    "codelyzer": "6.0.1",
-    "jasmine-core": "3.6.0",
-    "jasmine-spec-reporter": "6.0.0",
-    "karma": "5.2.3",
-    "karma-chrome-launcher": "3.1.0",
+    "@types/jasmine": "5.1.4",
+    "@types/jquery": "3.5.30",
+    "@types/node": "20.14.2",
+    "codelyzer": "6.0.2",
+    "jasmine-core": "5.1.2",
+    "jasmine-spec-reporter": "7.0.0",
+    "karma": "6.4.3",
+    "karma-chrome-launcher": "3.2.0",
     "karma-cli": "2.0.0",
     "karma-coverage-istanbul-reporter": "3.0.3",
-    "karma-jasmine": "4.0.1",
-    "karma-jasmine-html-reporter": "1.5.4",
+    "karma-jasmine": "5.1.0",
+    "karma-jasmine-html-reporter": "2.1.0",
     "protractor": "7.0.0",
-    "ts-node": "9.0.0",
+    "ts-node": "10.9.2",
     "tslint": "6.1.3",
-    "typescript": "4.0.3"
+    "typescript": "5.4.5"
   }
 }
