@@ -1,9 +1,9 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
 import { TranslateLoader } from '@ngx-translate/core';
 import { environment } from 'environments/environment';
-import 'rxjs/Rx';
+import { Subscription } from 'rxjs';
+import { Observable } from 'rxjs';
 
 @Injectable()
 
