@@ -1,16 +1,16 @@
 /* tslint:disable:no-unused-variable */
 
-import { TestBed, async, inject } from '@angular/core/testing';
-import { GroupService } from './group.service';
+import { TestBed, waitForAsync, inject } from '@angular/core/testing';
+import { GroupService } from './Group.service';
 
 describe('Service: Group', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [GroupService]
+    beforeEach(() => {
+        TestBed.configureTestingModule({
+            providers: [GroupService]
+        });
     });
-  });
 
-  it('should ...', inject([GroupService], (service: GroupService) => {
-    expect(service).toBeTruthy();
-  }));
+    it('should ...', inject([GroupService], (service: GroupService) => {
+        expect(service).toBeTruthy();
+    }));
 });
