@@ -3,14 +3,14 @@ import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatSort } from '@angular/material/sort';
 import { MatTableDataSource } from '@angular/material/table';
-import { LookUp } from 'app/core/models/lookUp';
+import { LookUp } from 'app/core/models/LookUp';
 import { AlertifyService } from 'app/core/services/alertify.service';
-import { LookUpService } from 'app/core/services/lookUp.service';
+import { LookUpService } from 'app/core/services/LookUp.service';
 import { environment } from 'environments/environment';
 import { IDropdownSettings } from 'ng-multiselect-dropdown';
 import { AuthService } from '../login/services/auth.service';
-import { Group } from './models/group';
-import { GroupService } from './services/group.service';
+import { Group } from './Models/Group';
+import { GroupService } from './Services/Group.service';
 
 
 declare var jQuery: any;
