@@ -4,10 +4,10 @@ import { MatPaginator } from '@angular/material/paginator';
 import { MatSort } from '@angular/material/sort';
 import { MatTableDataSource } from '@angular/material/table';
 import { AlertifyService } from 'app/core/services/alertify.service';
-import { LookUpService } from 'app/core/services/lookUp.service';
+import { LookUpService } from 'app/core/services/LookUp.service';
 import { AuthService } from '../login/services/auth.service';
-import { Language } from './Models/Language';
-import { LanguageService } from './Services/Language.service';
+import { Language } from './models/language';
+import { LanguageService } from './services/language.service';
 
 declare var jQuery: any;
 
