@@ -1,9 +1,9 @@
 import { HttpClient } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
 import { TranslateService } from '@ngx-translate/core';
-import { LookUp } from 'app/core/models/lookUp';
+import { LookUp } from 'app/core/models/LookUp';
 import { LocalStorageService } from 'app/core/services/local-storage.service';
-import { LookUpService } from 'app/core/services/lookUp.service';
+import { LookUpService } from 'app/core/services/LookUp.service';
 import { environment } from 'environments/environment';
 import { LoginUser } from './model/login-user';
 import { AuthService } from './services/auth.service';
