@@ -1,30 +1,30 @@
 ﻿import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import {environment} from "environments/environment"
+import { environment } from "environments/environment"
 import { Observable } from 'rxjs';
-import { OperationClaim } from '../Models/OperationClaim';
+import { OperationClaim } from '../models/operationclaim';
 
 @Injectable({
-  providedIn: 'root'
+    providedIn: 'root'
 })
 export class OperationClaimService {
 
-  constructor(private readonly _httpClient: HttpClient) { }
+    constructor(private readonly _httpClient: HttpClient) { }
 
 
-  getOperationClaimList(): Observable<OperationClaim[]> {
+    getOperationClaimList(): Observable<OperationClaim[]> {
 
-    return this._httpClient.get<OperationClaim[]>(environment.getApiUrl + '/operation-claims/')
-  }
+        return this._httpClient.get<OperationClaim[]>(environment.getApiUrl + '/operation-claims/')
+    }
 
-  getOperationClaim(id: number): Observable<OperationClaim> {
-    return this._httpClient.get<OperationClaim>(environment.getApiUrl  + `/operation-claims/${id}`)
-  }
+    getOperationClaim(id: number): Observable<OperationClaim> {
+        return this._httpClient.get<OperationClaim>(environment.getApiUrl + `/operation-claims/${id}`)
+    }
 
-  updateOperationClaim(operationClaim: OperationClaim): Observable<any> {
-    return this._httpClient.put(environment.getApiUrl  + "/operation-claims/", operationClaim, { responseType: 'text' });
+    updateOperationClaim(operationClaim: OperationClaim): Observable<any> {
+        return this._httpClient.put(environment.getApiUrl + "/operation-claims/", operationClaim, { responseType: 'text' });
 
-  }
+    }
 
 
 }
