@@ -2,7 +2,7 @@
 import { Injectable } from '@angular/core';
 import {environment} from 'environments/environment';
 import { Observable } from 'rxjs';
-import { Translate } from '../Models/Translate';
+import { Translate } from '../models/translate';
 
 @Injectable({
   providedIn: 'root'
