@@ -1,16 +1,16 @@
 /* tslint:disable:no-unused-variable */
 
-import { TestBed, async, inject } from '@angular/core/testing';
+import { TestBed, waitForAsync, inject } from '@angular/core/testing';
 import { AuthInterceptorService } from './auth-interceptor.service';
 
 describe('Service: AuthInterceptor', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [AuthInterceptorService]
+    beforeEach(() => {
+        TestBed.configureTestingModule({
+            providers: [AuthInterceptorService]
+        });
     });
-  });
 
-  it('should ...', inject([AuthInterceptorService], (service: AuthInterceptorService) => {
-    expect(service).toBeTruthy();
-  }));
+    it('should ...', inject([AuthInterceptorService], (service: AuthInterceptorService) => {
+        expect(service).toBeTruthy();
+    }));
 });
