@@ -2,7 +2,7 @@ import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { environment } from 'environments/environment';
 import { Observable } from 'rxjs';
-import { LookUp } from '../models/lookUp';
+import { LookUp } from '../models/LookUp';
 
 
 @Injectable({
