@@ -1,16 +1,16 @@
 /* tslint:disable:no-unused-variable */
 
-import { TestBed, async, inject } from '@angular/core/testing';
-import { TranslationService } from './translation.service';
+import { TestBed, inject } from '@angular/core/testing';
+import { TranslationService } from './Translation.service';
 
 describe('Service: Translation', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [TranslationService]
+    beforeEach(() => {
+        TestBed.configureTestingModule({
+            providers: [TranslationService]
+        });
     });
-  });
 
-  it('should ...', inject([TranslationService], (service: TranslationService) => {
-    expect(service).toBeTruthy();
-  }));
+    it('should ...', inject([TranslationService], (service: TranslationService) => {
+        expect(service).toBeTruthy();
+    }));
 });
